Added total mass computations for binary quadratic forms (from 2011).
This code supports the paper 

      Hanke -- "The Structure of Masses of rank $n$ Quadratic Lattices 
	         of varying determinant over number fields."

available at

      http://arxiv.org/abs/1108.3580

and computes the A_p and B_p quantities by summing the contributions
of each of the 2 Jordan Block decompositions (for p>2) or the 5 Partial
Local Genus Symbol contributions (for p=2).  It also computes the number of 
automorphisms and proper autmorphisms of binary quadratic forms quickly, 
to make the total masses of given determinant.